03-demos/vogue-concierge/vogue_concierge_agent/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.94.0
anyio==4.13.0
argparse==1.4.0
attrs==26.1.0
authlib==1.6.10
banks==2.4.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
colorama==0.4.6
cryptography==46.0.7
dataclasses-json==0.6.7
defusedxml==0.7.1
deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
docker==7.1.0
docstring-parser==0.17.0
durationpy==0.10
fastapi==0.135.3
fastuuid==0.14.0
filelock==3.25.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.3.0
google-adk==1.29.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-cloud-aiplatform==1.147.0
google-cloud-appengine-logging==1.9.0
google-cloud-audit-log==0.5.0
google-cloud-bigquery==3.41.0
google-cloud-bigquery-storage==2.37.0
google-cloud-bigtable==2.36.0
google-cloud-core==2.5.1
google-cloud-dataplex==2.18.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.22.0
google-cloud-logging==3.15.0
google-cloud-monitoring==2.30.0
google-cloud-pubsub==2.37.0
google-cloud-resource-manager==1.17.0
google-cloud-secret-manager==2.27.0
google-cloud-spanner==3.64.0
google-cloud-speech==2.38.0
google-cloud-storage==3.10.1
google-cloud-trace==1.19.0
google-crc32c==1.8.0
google-genai==1.72.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
graphviz==0.21
greenlet==3.4.0
griffe==2.0.2
griffecli==2.0.2
griffelib==2.0.2
grpc-google-iam-v1==0.14.4
grpc-interceptor==0.15.4
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
hf-xet==1.4.3
honcho==2.0.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.10.1
idna==3.11
importlib-metadata==8.7.1
jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
k8s-agent-sandbox==0.3.10
kubernetes==35.0.0
langchain-core==1.2.28
langgraph==0.4.7
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.30
litellm==1.82.6
llama-index-core==0.14.20
llama-index-embeddings-google-genai==0.5.0
llama-index-instrumentation==0.5.0
llama-index-readers-file==0.6.0
llama-index-workflows==2.17.3
lxml==6.0.4
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.2
mcp==1.27.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
oauthlib==3.3.1
openai==2.31.0
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pillow==12.2.0
platformdirs==4.9.6
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic-core==2.46.0
pydantic-settings==2.13.1
pygments==2.20.0
pyjwt==2.12.1
pyopenssl==26.0.0
pyparsing==3.3.2
pypdf==6.10.0
pypika==0.51.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.26
pytz==2026.1.post1
pyyaml==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sqlalchemy==2.0.49
sqlalchemy-spanner==1.17.3
sqlparse==0.5.5
sse-starlette==3.3.4
starlette==0.52.1
striprtf==0.0.26
tenacity==9.1.4
tiktoken==0.12.0
tinytag==2.2.1
tokenizers==0.22.2
toolbox-adk==1.0.0
toolbox-core==1.0.0
tqdm==4.67.3
typer==0.24.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.44.0
watchdog==6.0.0
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
